chore: refactor ios in-app classes and fix naming (#184)
