+++ ios/Classes/CustomerIOPlugin.h
@@ -1,0 +1,4 @@
+#import <Flutter/Flutter.h>
+
+@interface CustomerIOPlugin : NSObject<FlutterPlugin>
+@end
