--- ios/Classes/CustomerIoPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface CustomerIoPlugin : NSObject<FlutterPlugin>
-@end
